Move abicalldata to own package
