+++ lib/abicalldata/calldata_test.go
@@ -1,4 +1,4 @@
-package malleable
+package abicalldata
 
 import (
 	"math"
